fix: remove v1 functions, delete class based kura and migrate v1 orchestration into individual modules
